refactor: migrate scriptpreview
